add seSuperpone en reserva
